Added Q&A container component
