+++ package.json
@@ -8,7 +8,8 @@
     "server-dev": "webpack serve",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
+    "test:coverage": "jest --coverage",
+    "test:watch-alex": "jest --watch --testPathPattern='testsQA'"
   },
   "repository": {
     "type": "git",
