+++ src/components/qAndA/testsQA/Upvote.test.js
